+++ scripts/seed-npc-reviews-and-intel.ts
@@ -1,4 +1,4 @@
-#!/usr/bin/env ts-node
+#!/usr/bin/env node
 
 /**
  * NPC Reviews and Intel Seeding Script